Optimized buffer handling for raw RGB lcds.
Previously, we would create a buffer in memory for the current scene, then push that whole buffer to the esp-idf library to render.

This was inefficient as the size of the framebuffer requires holding it in PSRAM, and thus the present frame operation required a PSRAM to PSRAM copy that's slow and unnecessary.

The fix for this was to manage the active and drawing frame buffers manually, and manage the RGB lcd's bounce buffer manually. This means less copying per frame and thus improved performance.