Clear operations before the next get call.
Not clearing operations created bugs where unreadable SPI transactions would be mis-interpreted as the previous packet. That would get really buggy when byte buffers were being read and were modified by the reading of the stream.

This ensures that not changing an operation would have an operation of type zero, and thus no action would be taken.